Add `--glob` named argument to ls.
Description:
The `--glob` argument controls the interpretation of the user-given
pattern as a first step toward resolving #10364, which requests a
similar argument to all commands which glob by default.

User-Facing Changes:
Users can now give the `--glob` to `ls` in case they do not want their
expression to be interpreted as a glob. If they choose not to use this
argument then they will experience no changes.

Test Changes:
Two existing tests were updated to support this change.

The `flag_completions` test checks that all possible flags are
displayed to the user, and `ls` happens to be the command used for
suggestions. The `--glob` and `-g` flags were added to the expected
output.

The `lists_regular_files_in_special_folder` test was updated with 4 new
cases to confirm that file with square brackets are found when
`--glob false` is given, but not when `--glob true` is given.